Change ktor client to okhttp
Kotless is not compatible with ktor client.
It has NoClassDefFoundError everywhere.

Used OkHttp for the sole reason to make it
straightforward and easier since manual logic
was already there and adding Retrofit isn't
worth it for now because I don't foresee
this to add more endpoints that Retrofit
easily supports.