Fix reattach bug
State does not flow to the stream when a new container is attached.
This is because of scanReduce which only executes when there is an
accumulator. When you are re-attaching a Scene, it will only
emit the latest state and will not have an accumulator. This
solution let's us handle the reduction of state manually.

On an ideal scenario this logic should be in the ViewModel.